Add a long global doc-id space to composite readers
A composite/directory reader can now address more than Integer.MAX_VALUE
documents. Add IndexReader#totalMaxDoc()/totalNumDocs() (long) and widen
LeafReaderContext#docBase and the BaseCompositeReader starts[] to long.
The legacy int maxDoc()/numDocs() now throw ArithmeticException on a
composite reader past 2^31. Leaf readers and codec APIs are untouched.

The composite-as-one-leaf helpers (MultiBits, MultiTerms, MultiDocValues)
address the doc space through int APIs (Bits, DocIdSetIterator) and cannot
represent it beyond 2^31; they now fail fast at construction with a clear
message pointing to per-leaf access, via ReaderUtil#ensureIntDocIdSpace.
Add long global-id overloads to StoredFields#document and TermVectors#get.

+++ lucene/core/src/java/org/apache/lucene/index/CheckIndex.java
@@ -715,8 +715,8 @@ public final class CheckIndex implements Closeable {
     }
 
     if (infoStream != null) {
-      int maxDoc = 0;
-      int delCount = 0;
+      long maxDoc = 0;
+      long delCount = 0;
       for (SegmentCommitInfo info : lastCommit) {
         maxDoc += info.info.maxDoc();
         delCount += info.getDelCount();
